Update Server.java
Added readNameFile to read mainFile containing the usernames and passwords for everyone. Added variables "usernames", "passwords" and "accounts" to keep track of information during runtime. Added mainFile to be set as the main file for the server. Added basic signIn method. Added addAccount to add accounts to mainFile. Added two writeToFiles, one to add a single line to the end, and one to rewrite entire files. DeleteConvo, createConvo, and sendMessage will be done once we determine how we are actually sending messages.